[FIX] web: restore BS4 behaviour for setting element content
Fix some cases when we want to render more than one node on the root
level.

In BS4 the jQuery was using all node but in the jQuery compatibility
layer in BS5 it takes only the first node if the jQuery Element is used
in the function. This hack try to restore the old behaviour wrapping all
element in a DIV element (to be single root node).

PS: this can break some querySelector and some CSS selector if they are
too fragile.

e.g. of element that break before this hack:
Popover content for PaymentPopOver XML in ShowPaymentLineWidget