[FIX] mail: properly place dropdown in Kanban Activities When viewing the Kanban Activities (in the CRM module), when we have enough entries to require the scrolling, dropdowns are not displayed at the correct location. Furthermore, when opening dropdowns on the far right of the screen, part of them are rendered outside the screen.
Step to reproduce the issues:
1. Install the CRM module
2. Create many opportunities with an activity on each (at least 10)
3. Go to CRM and on the Pipeline view, select the "View Activity" layout
4. Scroll down to the last lines and click on an entry
You will see that the dropdown is not displayed at the expected location

Solution: The issue was that the `boundary` of the dropdown was not precise
enough (included since PR [1]). Instead of just using the `viewport`, we
can use dynamic query and target the HTML element containing the scrolling
view. To cope with the overflowing element (out of the screen), we can
dynamically use the `dropdown-menu-right` property based on the position and
size of the element.

[1]: https://github.com/odoo/odoo/pull/72441

opw-2827873