[FIX] various: fix traceback in survey with demo data
Currently, in demo data of some surveys, there are records of 'survey.question'
having 'is_page' field value as true and 'question_type' is not defined. When
question type is not defined, default value 'simple_choice' is assigned to the
'question_type' field. Now when we create a live session of that survey and
goes to the page, it tries to read the data for bar chart but data is not
available (undefined) and throws the traceback. This happens  because
question types ('simple_choice', 'multiple_choice') has bar chart.
Previous versions set the  default 'question_type' value to 'text_box',
so no traceback is triggered.

In this commit, we are setting the 'question_type' field value to 'False' in the
demo data record whose 'is_page' field value is 'True'.

taskID- 2841582