[IMP] survey: allow matrix answers to trigger conditional questions
PURPOSE:
Allow users to display questions based on answers provided through a
matrix-type question.

When checking the "is conditional" checkbox in the options of a question,
The user can now select a matrix-type question as triggering question. When
such a question is selected, the user can choose the row and the answer
of the matrix question that will trigger the question.

To be able to handle these matrices row's answers, selected answers are now
represented by a pair (suggested answer, matrix row), where the matrix row
is empty for non-matrix-type questions, when fetching data related to
conditional questions.

Task-2846710