[FIX] web, partner_autocomplete: fix placeholder inside widgets When multiple structured widgets (e.g. those creating DIVs on top of the usual input [1]) for the same field (e.g. through the invisible attribute) are created, the placeholder is not properly transmitted to the lower level `input` (which is inside the div).
Step to reproduce the issue (in v15):
1. Install Contact module
2. Go to Settings > Contacts and activate "Partner Autocomplete"
3. Go on the form to create a new Contact and before inputting a name, switch
between company and individual (with the radio button)
You will see that the placeholder never changes, while it should state
"Lumber Inc" when ticking Company

Solution: The attributes passed onto a widget are given by the state of the
renderer, which is stored as a dictionnary (field_name: attributes).
Consequently, when you have multiple fields with the same name (e.g. when you
use multiple fields with `invisible` attributes to control which field is shown
), the information are overwritten in the dictionary. In the other hand, when
creating a widget, the main HTML element of it is post processed by the
renderer to add some properties of the XML node directly (e.g. placeholder,
style, etc. cfr [2]). However, some widget creates a structure (e.g. a div
wrapping an input) and the placeholder of the inner elements are defined by the
attributes of the widget (thus, passed by the state of the renderer, not the
node, which means this information may not be linked to the XML node directly,
as information may be overriden in the state).

[1]: https://github.com/odoo/odoo/blob/c8c3aa96e18827c3c032540b92e95b27bbed0fc3/addons/partner_autocomplete/static/src/js/partner_autocomplete_fieldchar.js#L99
[2]: https://github.com/odoo/odoo/blob/c8c3aa96e18827c3c032540b92e95b27bbed0fc3/addons/web/static/src/js/views/basic/basic_renderer.js#L466

opw-2801527