[IMP] knowledge: update view invisible attributes for admins
Currently, the buttons allowing the user to move an article, archive an
article, invite a new member, etc are hidden if the user does not have
"write" permission on the article. However, the administrators should
always be able to perform those actions even if they do not have "write"
permission. This commit will update the `invisible` attributes of some
elements to add an exception for the administrators.

task-2852905