[FIX] knowledge: fix auto-save issue
When the user edits the document and quits the view, the user changes will
be discarded unless the user unfocuses the editor beforehand. This commit
will fix the auto-save issue by reducing the debouncing interval from
1000000000ms (default value) to 500ms. The editor should now do intermediate
saves when the user stops editing the document for 500ms. This change will
mitigate the losses if the user closes the tab or reloads the page.

When the `beforeLeave` event is triggered, the controller will now commit
all changes before saving them. This will guarantee that the records
associated with the debounced fields will be marked as "dirty" and that
the changes will be persisted in db.

task-2857029