[REF] mass_mailing: rename hash methods
In this commit we rename methods generating hashes (for recipients and report
management). Otherwise it is hard to distinguish them as naming is not really
convenient to use.

Hash for mail tracking is also moved in its own method to avoid having the
same code twice in codebase. Better delegate it to a tool sub method.

Task-2669037