[IMP] mrp: improve _compute_quantities_dict performance
We can iterate through boms instead of products to avoid computing the same quantities multiple times, for example in the case of variants which don't have different boms from the template. For each bom kit we can compute the quantities once and apply them for all relevant products at the end.