[IMP] knowledge: clean up the permission panel
This commit aims to improve the code quality and fix some issues with
the permission panel of Knowledge.

- The loader icon of the permission panel will now be displayed while the
first rpc call loading the data is pending.
- When the internal permission of an article changes or the permission of
the user changes, the article can move from one category to another
(e.g: from "private" to "workspace"). To reload the tree, we previously
reload the whole view which was quite expensive. The view will now be
reloaded partially (see: the `reload_view` and `reload_tree` events).
- When the user moves an article with a drag-and-drop or uses the "move to"
modal, the permission of the user on the article can change as the
permissions can be inherited from its parent. We will therefore reload
the permission panel when the user moves an article.
- When the user is about to lose access or the "write" permission on the
article, we will save the document before applying the permission changes.
That way, the user will not lose their changes.
- The links to the "based on" articles of the permission panel was not
working as the name of the action opening the article was incorrect.

task-2852905