[IMP] base: unform "groups" in back-end view
In a back-end view:
 - If a field has the `groups` attribute set and the user is not part of
   the groups, the field is removed, completely, from the view.
 - If a node has the `groups` attribute set and the user is not part of
   the groups, the node is made invisible.

The goal of this revision is to make the behavior of the `groups`
attribute on the field and on the node uniform:
In both case, remove the node from the view.

In front-end view, this is the case,
if a node has a "groups" or "t-groups" set and the user
is not member of the group, the node is removed from the view.

So it's 2/3 cases doing the same behavior.
It's simpler to have only one uniform behavior for this attribute.
Simpler for developers and users to understand the behavior in all
cases.

In addition, this will help to use the qweb engine to compile
back-end views, if we do so in the future.