[FIX] website_mass_mailing: hide thanks btn of newsletter before saving
Before this commit the "Thanks" button of the newsletter block could be
saved as visible. This meant that a visitor who did not subscribe to the
newsletter would see the "Thanks" button on loading before it
disappeared. This commit solves this problem by hiding the button before
saving.

Steps to reproduce the fixed bug:
 - Drop a newsletter block.
 - Save and register to the newsletter.
 - Edit
 - Change anything in the newsletter block.
 - Save and Logout.
 - Now the Thanks button appears and disappears when the page is loaded

opw-2802139