[IMP] survey: improve survey samples accessibility
PURPOSE:

This commit improves the accessibility of the survey samples when no records
are shown.

SPECS

Remove the modal used to show the samples and show them on the empty screen
instead.

When no survey exists, the user gets the proposition of trying survey samples.
Previously, the user had to click on a button to open a modal showing the
different samples. Now, the samples are shown in the empty screen directly.

Task-2794884