[IMP] survey: revamp onboarding data & test surveys integration
PUPRPOSE:

This commit revamps the onboarding data and slightly modifies the integration
of test surveys and related filters.

SPECS

1. Revamp sample surveys data

Considering this is the data we want to push forward, it needs to be as
relevant as possible while being "light" for the db.

2. Consider completed test surveys as completed surveys

Most people expect to see the test the just did counted in the completed
surveys counter in the survey view. The filter excluding the tests surveys has
therefore also been removed from the participation view.

3. Rename the filters related to the test surveys

"Tests Only" and "Exclude Tests" are more straight to the point than
"Test Entries" and "Exclude Test Entries".

Task-2794884