[IMP] survey: implement a tour
PURPOSE:

This commit adds a tour that is started when the survey module is
installed, and makes small changes to be able to implement the tour.

SPECS

1. Implementation of the tour

The tour suggests to the user to try and test a sample survey, and shows how
to check the results gathered for this survey.

2. Changed the target of the "test" button to "self"

Since the test action is used during the tour, the test page needs to be opened
in the same tab so that the tour can continue.

Task-2794884