[FIX] web_editor: compare size CSS properties
When CSS properties are updated, a check is made to verify whether the
new value is different from the old one. This process sometimes involves
unit conversions.
Since [1] when a CSS property is updated, it is also set before checking
in order to verify whether it should be set as `!important`.
When the CSS property is a size, such as the `background-size` CSS
property, it is made of two dimensions, but the verification does not
handle that situation: it handles `15px auto` as being a single number
with the unit `px auto`, which is wrong.
The full comparison is done because the strings do not fully match:
when a size property is set to a measure first then `auto`, it is
returned without `auto` when reading the computed style.
E.g. after setting `15px auto`, it reads as `15px`.

This commit adapts the CSS property comparison mechanism into splitting
the size properties into their parts and comparing each part separately.

Steps to reproduce:
- Drop a "Banner" block.
- Set "Position" as "Repeat Pattern".
- Type "5" in the "Width".
=> Did raise an error.

[1]: https://github.com/odoo/odoo/commit/d3c3dab8950abc25b29937605091d8ce32305fa4

task-2853161