[IMP] project : add and modify commands in the palette
- Add 'Move to next stage':

About 'Move to stage...' not being in the command palette,
Stage and Personal Stage both have a status bar widget.
Yet, when this widget creates a command in the registry,
it stores it at the same key. So it overrides the last saved.
This is the reason why this command was not available.
An improvement to come in the framework js aims to not create
commands for invisible fields. As, in our case, only one will be visible
at a time, the conflict will be resolved.
We added a 'Move to next Stage/Personal User Stage command'.

- Change the 'Set priority...' command
to 'Set priority as Low/High' :

Depending on the priority of the ticket, the command will display
'Set priority as [the other priority]'.

- Transform the 'Set kanban state...' command into three commands

Only two of them are now displayed,
depending on the current kanban state of the task.

task-2857447