[FIX] mail: fix discuss container removing discussView
The DiscussContainer clears the discussView when onWillUnmount is called.
The issue occurs when the discuss action is called while it is already active.
Indeed, the setup of the newest DiscussContainer will update the discussView
then the unmount of the oldest one will clear it. This results in a potentially
missing discussView.