[FIX] account: fix expense total computation
Current behavior :
The amount shown for an expense is equals to 0 and not the real price

Steps :
- Install Expenses, Accounting
- Create an expense and finish the flow (register payment included)
- In Accounting app, go to Vendors > Payments
- In this view, amount is equals to 0 but the amount is correct in individual view

Note :
Since this commit [1], the condition has changed and allows more cases.
In this case, we now enter this condition [2].

This causes problems when recording payments for our expense.
Indeed, the 2 lines (credit + debit) enter this condition [3] .

As this is the first one, it doesn't go any further and results in a total of 0
(the 2 lines cancel each other out). However, the debit line should be treated
by this block [4] which is not accessible since we stop at the first condition.

[1] : b7b9b18ac1d93b5f7de088a4c92f19488c05fc56
[2] : https://github.com/odoo/odoo/blob/d6ca59ed9bf6877ee6b1b311223472adfa4db549/addons/account/models/account_move.py#L1419
[3] : https://github.com/odoo/odoo/blob/d6ca59ed9bf6877ee6b1b311223472adfa4db549/addons/account/models/account_move.py#L1422
[4] : https://github.com/odoo/odoo/blob/d6ca59ed9bf6877ee6b1b311223472adfa4db549/addons/account/models/account_move.py#L1434-L1438

OPW-2838105